add: SpatialPromptAnalyzer
