+++ hemm/metrics/spatial_alignment/__init__.py
@@ -1,0 +1,3 @@
+from .analyze_prompt import SpatialPromptAnalyzer
+
+__all__ = ["SpatialPromptAnalyzer"]
